+++ trader/infrastructure/persistence/db.go
@@ -9,3 +9,19 @@ type DB interface {
 	Query(query string, args ...interface{}) (*sql.Rows, error)
 	QueryRow(query string, args ...interface{}) *sql.Row
 }
+
+func NewMySQLTransaction(dsn string) *sql.Tx {
+	db, err := sql.Open("mysql", dsn)
+	if err != nil {
+		panic(err.Error())
+	}
+
+	err = db.Ping()
+	if err != nil {
+		panic(err.Error())
+	}
+
+	tx, _ := db.Begin()
+
+	return tx
+}
